move set initial state to thunks test, update snapshots
