+++ src/modules/blocks/event-datetime/block.js
@@ -86,14 +86,13 @@ class EventDateTime extends Component {
 		setInitialState: PropTypes.func,
 		setCost: PropTypes.func,
 		openDashboardDateTime: PropTypes.func,
-		setStart: PropTypes.func,
-		setEnd: PropTypes.func,
-		setMultiDay: PropTypes.func,
+		setStartTime: PropTypes.func,
+		setEndTime: PropTypes.func,
+		setMultiDayThunk: PropTypes.func,
+		setDates: PropTypes.func,
 		setTimeZone: PropTypes.func,
 		setSeparatorTime: PropTypes.func,
 		setSeparatorDate: PropTypes.func,
-		setStartTime: PropTypes.func,
-		setEndTime: PropTypes.func,
 		closeDashboardDateTime: PropTypes.func,
 		setVisibleMonth: PropTypes.func,
 		visibleMonth: PropTypes.instanceOf( Date ),
