+++ src/modules/data/blocks/datetime/__tests__/actions.test.js
@@ -1,15 +1,7 @@
 /**
- * External dependencies
- */
-import configureStore from 'redux-mock-store';
-import thunk from 'redux-thunk';
-/**
  * Internal dependencies
  */
 import { actions } from 'data/blocks/datetime';
-
-const middlewares = [ thunk ];
-const mockStore = configureStore( middlewares );
 
 describe( '[STORE] - Datetime actions', () => {
 	test( 'Action to set the start time', () => {
@@ -40,31 +32,5 @@ describe( '[STORE] - Datetime actions', () => {
 	test( 'Action to set the multi day', () => {
 		expect( actions.setMultiDay( true ) ).toMatchSnapshot();
 		expect( actions.setMultiDay( false ) ).toMatchSnapshot();
-	} );
-} );
-
-describe( '[STORE] - Datetime thunk actions', () => {
-	test( 'Initial set of state', () => {
-		const store = mockStore( {} );
-
-		const attributes = {
-			start: 'June 5, 2018 5:00 pm',
-			end: 'June 5, 2018 5:30 pm',
-			dateTimeSeparator: ' @ ',
-			timeRangeSeparator: ' - ',
-			allDay: false,
-			multiDay: false,
-			timezone: 'UTC',
-		};
-
-		const get = jest.fn( ( key ) => {
-			return attributes[ key ];
-		} );
-
-		store.dispatch( actions.setInitialState( { get, attributes } ) );
-
-		expect( get ).toHaveBeenCalled();
-		expect( get ).toHaveBeenCalledTimes( 2 );
-		expect( store.getActions() ).toMatchSnapshot();
 	} );
 } );
